[mob][photos] small refactor
