+++ mobile/lib/models/ml/ml_versions.dart
@@ -2,3 +2,6 @@ const faceMlVersion = 1;
 const clipMlVersion = 1;
 const clusterMlVersion = 1;
 const minimumClusterSize = 2;
+
+const embeddingFetchLimit = 200;
+const fileDownloadMlLimit = 10;
