+++ mobile/lib/services/machine_learning/face_ml/face_recognition_service.dart
@@ -5,17 +5,8 @@ import "dart:ui" show Image;
 
 import "package:logging/logging.dart";
 import "package:photos/core/event_bus.dart";
-import "package:photos/db/ml/clip_db.dart";
-import "package:photos/db/ml/db.dart";
 import "package:photos/events/diff_sync_complete_event.dart";
 import "package:photos/events/people_changed_event.dart";
-import "package:photos/extensions/list.dart";
-import "package:photos/models/ml/clip.dart";
-import "package:photos/models/ml/face/face.dart";
-import "package:photos/models/ml/ml_versions.dart";
-import "package:photos/service_locator.dart";
-import "package:photos/services/filedata/filedata_service.dart";
-import "package:photos/services/filedata/model/file_data.dart";
 import "package:photos/services/machine_learning/face_ml/face_detection/detection.dart";
 import "package:photos/services/machine_learning/face_ml/face_detection/face_detection_service.dart";
 import "package:photos/services/machine_learning/face_ml/face_embedding/face_embedding_service.dart";
@@ -23,7 +14,6 @@ import "package:photos/services/machine_learning/face_ml/person/person_service.d
 import "package:photos/services/machine_learning/ml_exceptions.dart";
 import "package:photos/services/machine_learning/ml_result.dart";
 import "package:photos/utils/image_ml_util.dart";
-import "package:photos/utils/ml_util.dart";
 
 class FaceRecognitionService {
   final _logger = Logger("FaceRecognitionService");
@@ -39,8 +29,6 @@ class FaceRecognitionService {
 
   bool _shouldSyncPeople = false;
   bool _isSyncing = false;
-
-  static const _embeddingFetchLimit = 200;
 
   Future<void> init() async {
     if (_isInitialized) {
@@ -78,101 +66,6 @@ class FaceRecognitionService {
       _shouldSyncPeople = false;
     }
     _isSyncing = false;
-  }
-
-  Stream<List<FileMLInstruction>> syncEmbeddings({
-    int yieldSize = 10,
-  }) async* {
-    final List<FileMLInstruction> filesToIndex = await getFilesForMlIndexing();
-    final List<List<FileMLInstruction>> chunks =
-        filesToIndex.chunks(_embeddingFetchLimit);
-    List<FileMLInstruction> batchToYield = [];
-
-    for (final chunk in chunks) {
-      if (!localSettings.remoteFetchEnabled) {
-        _logger.warning("remoteFetchEnabled is false, skiping embedding fetch");
-        yield chunk;
-      }
-      final Set<int> ids = {};
-      final Map<int, FileMLInstruction> pendingIndex = {};
-      for (final instruction in chunk) {
-        ids.add(instruction.file.uploadedFileID!);
-        pendingIndex[instruction.file.uploadedFileID!] = instruction;
-      }
-      _logger.info("fetching embeddings for ${ids.length} files");
-      final res = await FileDataService.instance.getFilesData(ids);
-      _logger.info("embeddingResponse ${res.debugLog()}");
-      final List<Face> faces = [];
-      final List<ClipEmbedding> clipEmbeddings = [];
-      for (FileDataEntity fileMl in res.data.values) {
-        final existingInstruction = pendingIndex[fileMl.fileID]!;
-        final facesFromRemoteEmbedding = _getFacesFromRemoteEmbedding(fileMl);
-        //Note: Always do null check, empty value means no face was found.
-        if (facesFromRemoteEmbedding != null) {
-          faces.addAll(facesFromRemoteEmbedding);
-          existingInstruction.shouldRunFaces = false;
-        }
-        if (fileMl.clipEmbedding != null &&
-            fileMl.clipEmbedding!.version >= clipMlVersion) {
-          clipEmbeddings.add(
-            ClipEmbedding(
-              fileID: fileMl.fileID,
-              embedding: fileMl.clipEmbedding!.embedding,
-              version: fileMl.clipEmbedding!.version,
-            ),
-          );
-          existingInstruction.shouldRunClip = false;
-        }
-        if (!existingInstruction.pendingML) {
-          pendingIndex.remove(fileMl.fileID);
-        } else {
-          existingInstruction.existingRemoteFileML = fileMl;
-          pendingIndex[fileMl.fileID] = existingInstruction;
-        }
-      }
-      for (final fileID in pendingIndex.keys) {
-        final instruction = pendingIndex[fileID]!;
-        if (instruction.pendingML) {
-          batchToYield.add(instruction);
-          if (batchToYield.length == yieldSize) {
-            _logger.info("queueing indexing for  $yieldSize");
-            yield batchToYield;
-            batchToYield = [];
-          }
-        }
-      }
-      await MLDataDB.instance.bulkInsertFaces(faces);
-      await MLDataDB.instance.putMany(clipEmbeddings);
-    }
-    // Yield any remaining instructions
-    if (batchToYield.isNotEmpty) {
-      _logger.info("queueing indexing for  ${batchToYield.length}");
-      yield batchToYield;
-    }
-  }
-
-  // Returns a list of faces from the given remote fileML. null if the version is less than the current version
-  // or if the remote faceEmbedding is null.
-  List<Face>? _getFacesFromRemoteEmbedding(FileDataEntity fileMl) {
-    final RemoteFaceEmbedding? remoteFaceEmbedding = fileMl.faceEmbedding;
-    if (shouldDiscardRemoteEmbedding(fileMl)) {
-      return null;
-    }
-    final List<Face> faces = [];
-    if (remoteFaceEmbedding!.faces.isEmpty) {
-      faces.add(
-        Face.empty(fileMl.fileID),
-      );
-    } else {
-      for (final f in remoteFaceEmbedding.faces) {
-        f.fileInfo = FileInfo(
-          imageHeight: remoteFaceEmbedding.height,
-          imageWidth: remoteFaceEmbedding.width,
-        );
-        faces.add(f);
-      }
-    }
-    return faces;
   }
 
   static Future<List<FaceResult>> runFacesPipeline(
